myapp/requirement.txt
python-dotenv
fire==0.4.0